Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==3.1.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==4.11.3
Flask-PyMongo==3.0.1
jaeger-client==4.8.0
prometheus-client==0.21.1
opentelemetry-api==1.31.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation==0.52b1
opentelemetry-sdk==1.31.1
prometheus-flask-exporter==0.23.2
flask-opentracing==2.0.0
requests==2.28.2
urllib3==1.26.15
six==1.16.0